+++ docs/src/markdown/changelog.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## 3.1.0
+## 4.0.0
 
 - **NEW**: Deprecated `WcMatch` class methods `kill` and `reset`. `WcMatch` should be broken with a simple `break` statement instead.
 - **NEW**: Add a new flag `MARK` to force `glob` to return directories with a trailing slash.
 - **NEW**: Add `MATCHBASE` that causes `glob` and `WcMatch`, when the pattern has no slashes in it, to seek for any file anywhere in the tree with a matching basename.
+- **NEW**: Exclusion patterns (enabled with `NEGATE`) now always enable `DOTALL` in the exclusion patterns. They also will match symlinks in `**` patterns. Only non `NEGATE` patterns that are paired with a `NEGATE` pattern are subject to symlinks and dot rules. Exclusion patterns themselves allow dots and symlinks to make filtering easier.
+- **NEW**: Exclusion patterns no longer provide a default regular pattern if one is not applied. Exclusion patterns are
+meant to filter the results of normal patterns. You can either use the `SPLIT` flag and provide a pattern with your default ('default_pattern|!exclusion'), or feed in a list of multiple patterns instead of a single string (`['default_pattern', '!exclusion']`). If you really need the old behavior, you can use the `NEGDEFAULT` flag which will provide a default of `**` which is subject to the `GLOBSTAR` flag. `NEGDEFAULT` will raise a deprecation warning and will be removed in the future.
+- **FIX**: An empty pattern in `glob` should not match slashes.
 
 ## 3.0.2
 
@@ -40,7 +44,7 @@
 ## 2.1.0
 
 - **NEW**: Deprecate `version` and `version_info` in favor of the more standard `__version__` and `__version_info__`.
-- **FIX**: Fix issue where negated patterns would trigger before end of path.
+- **FIX**: Fix issue where exclusion patterns would trigger before end of path.
 - **FIX**: Fix `GLOBSTAR` regular expression pattern issues.
 
 ## 2.0.3